Agregando y modificando components y in de controlador
