--- frontend/src/main/java/com/project/frontend/productsSystem/controllers/in/IGetAllOperation.java
@@ -1,9 +1,0 @@
-package com.project.frontend.productsSystem.controllers.in;
-
-import java.io.IOException;
-import java.util.List;
-import com.project.frontend.productsSystem.models.Product;
-
-public interface IGetAllOperation {
-     List<Product> getAllProducts() throws IOException;
-}
