--- frontend/src/main/java/com/project/frontend/productsSystem/controllers/in/IGetByFiltersOperation.java
@@ -1,9 +1,0 @@
-package com.project.frontend.productsSystem.controllers.in;
-
-import java.io.IOException;
-import java.util.List;
-import com.project.frontend.productsSystem.models.Product;
-
-public interface IGetByFiltersOperation {
-    List<Product> findByFilters(String nameProduct, String category, String nameSupplier) throws IOException;
-}
